refactor: clean up special casing and use ddof=1
Only with at least two historical measurements, the stddev can be
understood. Otherwise, adding the first measurement leads to having
a "0" stdddev for the next measurement. This is impossible to achieve.
Therefore use "n-1" for the divisor in the stddev calculation to keep
everything np.nan until there are at least two values.