feat: do not consider intermediate performance results from branches
If branches are merged with a merge commit, the intermediate results
from the features branches should not contribute to the history of
the main branch. More so, since many of the commits from the feature
branches might not even have a performance result. This would reduce
the depth of the history for the performance records.