feat: pass audit if there are too few measurements (#41)
In case that only a few measurements have been collected, the stddev
is not yet stable enough. Therefore, the audit should pass in those
cases.

Since the number of commits passed by `-n` only determines the
look-behind of commits scanned, and those commits are not guaranteed to
all have measurements, a second parameter --min-measurements is added.