degrees of freedom for variance is equal to N
For the rust implementation we did not repeat the bug from the python
implementation: ddof should be zero. The sample equals the population.
Therefore, no correction should be applied.

The first test was adapted as a single measurement no longer yields NaN
for the variance.

The last two tests were deleted, as they no longer add coverage
compared to the fixed, first test.

A new test for --min-measurements was added.